+++ azure/container-linux/kubernetes/workers.tf
@@ -9,7 +9,7 @@ module "workers" {
   security_group_id       = "${azurerm_network_security_group.worker.id}"
   backend_address_pool_id = "${azurerm_lb_backend_address_pool.worker.id}"
 
-  count    = "${var.worker_count}"
+  worker_count    = "${var.worker_count}"
   vm_type  = "${var.worker_type}"
   os_image = "${var.os_image}"
   priority = "${var.worker_priority}"