+++ azure/container-linux/kubernetes/workers/variables.tf
@@ -32,7 +32,7 @@ variable "backend_address_pool_id" {
 
 # instances
 
-variable "count" {
+variable "worker_count" {
   type        = "string"
   default     = "1"
   description = "Number of instances"