+++ google-cloud/container-linux/kubernetes/workers.tf
@@ -6,7 +6,7 @@ module "workers" {
   # GCE
   region       = "${var.region}"
   network      = "${google_compute_network.network.name}"
-  count        = "${var.worker_count}"
+  worker_count        = "${var.worker_count}"
   machine_type = "${var.worker_type}"
   os_image     = "${var.os_image}"
   disk_size    = "${var.disk_size}"